updated states file

+++ api/v1/views/states.py
@@ -1,56 +1,63 @@
 #!/usr/bin/python3
 '''Contains the states view for the API.'''
 
-from flask import Flask, jsonify, request, abort
-from api.v1.views import app_views
 from models import storage
 from models.state import State
+from api.v1.views import app_views
+from flask import jsonify, abort, request
+
 
 @app_views.route('/states', methods=['GET'], strict_slashes=False)
 def get_states():
-    states = [state.to_dict() for state in storage.all('State').values()]
-    return jsonify(states)
+    dict_ = []
+    for val in storage.all(State).values():
+        dict_.append(val.to_dict())
+    return jsonify(dict_)
 
-@app_views.route('/states/<state_id>', methods=['GET'], strict_slashes=False)
+
+@app_views.route('/states/<path:state_id>')
 def get_state(state_id):
-    '''Gets the state with the given id or all states.'''
-    state = storage.get('State', state_id)
+    state = storage.get(State, state_id)
     if state is None:
         abort(404)
     return jsonify(state.to_dict())
 
-@app_views.route('/states/<state_id>', methods=['DELETE'], strict_slashes=False)
+
+@app_views.route('/states/<path:state_id>', methods=['DELETE'], strict_slashes=False)
 def delete_state(state_id):
-    '''Removes a state with the given id.'''
-    state = storage.get('State', state_id)
+    if state_id is None:
+        abort(404)
+    state = storage.get(State, state_id)
     if state is None:
         abort(404)
     state.delete()
     storage.save()
-    return jsonify({}), 200
+    return jsonify({})
+
 
 @app_views.route('/states', methods=['POST'], strict_slashes=False)
-def create_state():
-    '''Adds a new state.'''
-    if not request.get_json():
-        abort(400, description="Not a JSON")
-    if 'name' not in request.get_json():
-        abort(400, description="Missing name")
-    state = State(**request.get_json())
-    state.save()
-    return jsonify(state.to_dict()), 201
+def post_state():
+    res = request.get_json()
+    if type(res) != dict:
+        return abort(400, {'message': 'Not a JSON'})
+    if 'name' not in res:
+        return abort(400, {'message': 'Missing name'})
+    new_state = State(**res)
+    new_state.save()
+    return jsonify(new_state.to_dict()), 201
 
-@app_views.route('/states/<state_id>', methods=['PUT'], strict_slashes=False)
-def update_state(state_id):
-    '''Updates the state with the given id.'''
-    state = storage.get('State', state_id)
+
+@app_views.route('/states/<path:state_id>', methods=['PUT'], strict_slashes=False)
+def put_state(state_id):
+    state = storage.get(State, state_id)
     if state is None:
         abort(404)
-    if not request.get_json(): 
-        abort(400, description="Not a JSON")
-    for key, value in request.get_json().items():
-        if key not in ['id', 'created_at', 'updated_at']:
+    res = request.get_json()
+    if type(res) != dict:
+        return abort(400, {'message': 'Not a JSON'})
+    for key, value in res.items():
+        if key not in ["id", "state_id", "created_at", "updated_at"]:
             setattr(state, key, value)
-    state.save()
+    storage.save()
     return jsonify(state.to_dict()), 200
